+++ packages/tautulli/lib/src/models/activity/session.dart
@@ -488,7 +488,7 @@ class TautulliSession {
     @JsonKey(name: 'deleted_user', toJson: TautulliUtilities.booleanToInteger, fromJson: TautulliUtilities.integerToBoolean)
     final bool deletedUser;
 
-    /// Allow guest?
+    /// Does the user have guest access to Tautulli?
     @JsonKey(name: 'allow_guest', toJson: TautulliUtilities.booleanToInteger, fromJson: TautulliUtilities.integerToBoolean)
     final bool allowGuest;
 
@@ -1079,8 +1079,8 @@ class TautulliSession {
     @override
     String toString() => json.encode(this.toJson());
 
-    /** JSON (De)Serialization Related */
-
+    /// Deserialize a JSON map to a [TautulliSession] object.
     factory TautulliSession.fromJson(Map<String, dynamic> json) => _$TautulliSessionFromJson(json);
+    /// Serialize a [TautulliSession] object to a JSON map.
     Map<String, dynamic> toJson() => _$TautulliSessionToJson(this);
 }